fix(dropdown): changed imports order
