+++ libs/dropdown/src/lib/dropdown/dropdown.component.ts
@@ -9,7 +9,20 @@ import {
 } from '@angular/core';
 import { ToBoolean } from '@spryker/utils';
 
-import { DropdownItem, Placement } from './types';
+export interface DropdownItem {
+  action: string;
+  title: string;
+  icon?: string;
+  disabled?: boolean;
+  subItems?: DropdownItem[];
+}
+export type Placement =
+  | 'bottomLeft'
+  | 'bottomCenter'
+  | 'bottomRight'
+  | 'topLeft'
+  | 'topCenter'
+  | 'topRight';
 
 @Component({
   selector: 'spy-dropdown',
